Flytter producer til egen mappe
