+++ apps/helsebro/src/main/kotlin/no/nav/helsearbeidsgiver/inntektsmelding/helsebro/App.kt
@@ -2,7 +2,7 @@ package no.nav.helsearbeidsgiver.inntektsmelding.helsebro
 
 import com.github.navikt.tbd_libs.rapids_and_rivers_api.RapidsConnection
 import no.nav.helse.rapids_rivers.RapidApplication
-import no.nav.helsearbeidsgiver.felles.rapidsrivers.pritopic.PriProducer
+import no.nav.helsearbeidsgiver.felles.kafka.Producer
 import no.nav.helsearbeidsgiver.utils.log.logger
 
 private val logger = "helsearbeidsgiver-im-helsebro".logger()
@@ -10,26 +10,26 @@ private val logger = "helsearbeidsgiver-im-helsebro".logger()
 fun main() {
     logger.info("im-helsebro er oppe og kjører!")
 
-    val priProducer = PriProducer()
+    val producer = Producer()
 
     RapidApplication
         .create(System.getenv())
-        .createHelsebroRivers(priProducer)
+        .createHelsebroRivers(producer)
         .start()
 
     logger.info("Nå dør jeg :(")
 }
 
-fun RapidsConnection.createHelsebroRivers(priProducer: PriProducer): RapidsConnection =
+fun RapidsConnection.createHelsebroRivers(producer: Producer): RapidsConnection =
     also {
         logger.info("Starter ${TrengerForespoerselRiver::class.simpleName}...")
-        TrengerForespoerselRiver(priProducer).connect(this)
+        TrengerForespoerselRiver(producer).connect(this)
 
         logger.info("Starter ${ForespoerselSvarRiver::class.simpleName}...")
         ForespoerselSvarRiver().connect(this)
 
         logger.info("Starter ${HentForespoerslerForVedtaksperiodeIdListeRiver::class.simpleName}...")
-        HentForespoerslerForVedtaksperiodeIdListeRiver(priProducer).connect(this)
+        HentForespoerslerForVedtaksperiodeIdListeRiver(producer).connect(this)
 
         logger.info("Starter ${VedtaksperiodeIdForespoerselSvarRiver::class.simpleName}...")
         VedtaksperiodeIdForespoerselSvarRiver().connect(this)
